Restructured lib with new data folder
