+++ src/app/merchandise/page.tsx
@@ -1,6 +1,6 @@
 import LoadingDots from '@/components/loading-dots';
 import Hero from '@/components/root-page/hero';
-import { fetchProducts } from '@/lib/api/products-data-server';
+import { fetchProducts } from '@/lib/data/services/api-product-service';
 import { fallbackDataManager } from '@/lib/mocks/fallback-data/fallback-data-manager';
 import { Product } from '@/lib/types/product';
 import { lazyMinLoadTime } from '@/lib/utils';
