+++ src/components/layout/pages-nav.tsx
@@ -7,8 +7,8 @@ import {
   NavigationMenuTrigger,
   navigationMenuTriggerStyle,
 } from '@/components/ui/navigation-menu';
-import { getCategories } from '@/lib/api/product-data-service';
 import { routes } from '@/lib/constants/routes';
+import { getCategories } from '@/lib/data/product-data-service';
 import Link from 'next/link';
 
 export async function PagesNav({ className }: { className?: string }) {
